Raspberry/requirements.txt
RPi.GPIO>=0.7.0; platform_machine == 'armv7l'  # Only install on Raspberry Pi
